Change grid min and max behavior to depend on 'as_midp' parameter
Grids have to play a double role in ODL, on the one hand as a pure set
as a collection of points, on the other hand as points in a tesselation,
or partition, of a spatial domain into small boxes. This behavior is
controlled by the `as_midp` parameter which becomes a state of the
grid object - `True` means the points are the midpoints of rectangular
cells, and the `min`, `max` and `cell_sizes` methods will take this
into account, while for `False`, the grid is a pure set without any
implied tesselation.

The main reason for this double behavior is that we do not need a
flexible approach for tesselations (yet) and want to avoid the
creation of an extra class just to change the behavior of three methods.

TODO: grid tests need to be fixed