Added override args for min and max to RegularGrid
Due to numerical errors, the calculated minimum and maximum of a
grid generated by the uniform_sampling function could lie outside
the sampled interval product, which caused
`intv_prod.contains_set(grid)` to evaluate to `False` for that
case. This in turn created lots of problems with checks down the
line.

Issues: #144