Enhance logic of dspace_type and add test for it
Before, `dspace_type` would try to infer the data space type solely
from the `field` attribute of the given `LinearSpace`. If (in the case
of `'numpy'` implementation) the field was `RealNumbers()`, it would
return `Rn`, for field `ComplexNumbers()` the result was `Cn`, and
if there was no `field`, the returned type would be `Ntuples`.

The problem with this logic occurs if one uses e.g. `dtype=int` in
`uniform_discr`. It would try to create an instance of the class returned
by `dspace_type` with `dtype=int`, which (now!) fails in many cases.

What one actually wants is that the data type determines if
`Rn/Cn` can be used or if one has to create the more general `Fn`.
This new logic is now implemented in `dspace_type` which takes an
additional `dtype` argument.

This function has a test now.