BUG: fix bug in test of newton solver.
This commit fix a bug in the test of the Newton solver. The bug was
due to radom input data (a quadratic proble x^THx + c^Tx, where c was
random), which caused the cg method, used to solve the linear system
Hx = -c, to crash some times. Now a constant c is used. See #36.