ENH: Rewrite vectorize decorator completely
The limitations of the old vectorizer became evident in #188 where
support for Numba would not work with the existing approach. This
commit replaces the old homebrew vectorizer with a vectorization
decorator based on numpy.vectorize.

The new 'vectorize' decorator can be used both with and without
arguments. An abstract class 'OptionalArgsDecorator' implementing
this pattern has been added.