ENH: improve thread and dtype handling
The number of threads in pyfftw_call is now set to the number of CPUs
with the help of the multiprocessing.cpu_count() function if the transform
size is at least 1000.

Available floating point data types and their complex counterparts are
determined from np.sctypes['float'] and np.result_type(dtype, 1j).

Further changes:
- Use utility conj_exponent from lp_discr
- Rename _pyfftw_check_in_out -> _pyfftw_check_args
- Allow file handle as [import,export]_wisdom parameter in pyfftw_call