ENH: add uniform_discr_frompartition and sequence_space
The new uniform_discr_frompartition factory is the new "most basic"
uniform discretization function, and uniform_discr_fromspace uses
it to avoid code duplication. A new helper default_dtype was added
to reduce duplication of code that needs to determine the default
data type for a certain combination of implementation and field.

A new simple factory sequence_space basically creates a uniform
discretization of the interval [0, n] with n subintervals, in order
to mimic a sequence space (with obvious generalization to higher
dimensions). The space uses the plain space functions of the underlying
data space, i.e. no weighting.