BUG: fix failing function discretization in Python 2
Since tuples can be compared to scalars in Python 2, expressions like
'tpl > 0' evaluate to True in Py2 for a non-empty tuple tpl. This is
triggered if a function using the expression 'x > 0' is fed with a
meshgrid tuple, which results in wrong scalar output instead of
vectorized evaluation of the expression with x[0], the first and only
entry in the tuple.
Even worse, indexing expressions like 'x[x > 0]' will automatically
trigger an IndexError for meshgrid tuple input, since 'x > 0'
evaluates to True, hence 1, and the index 1 is out of range for a
1-element tuple.
These situations are accounted for now in FunctionSetVector.__call__.

Additionally, meshgrids are now always tuples, and everything else
is regarded as an array. This allows array-like input to
vectorized functions.