ENH: add temporaries to FourierTransform parameters and factor out some code
Temporaries can now be given during init or created by a method. They are used
internally to avoid creation of new arrays. This is only relevant for mixed
data type transforms, not C2C. Not much speed gain probably, but perhaps the
transform becomes more memory-friendly. The temporaries are shared between a
transform and its inverse.