API/MAINT/ENH: Remove convex, concave, and smooth from functional.
This commit also adds the convex conjugate to all convex conjugate helper
classes, like ConvexConjugateTranslation etc.

It also removes 'to clever' checks for when to allow a call to the proximal in
FunctionalLeftScalarMult.

Moreover, it also checks if the input functional in TranslatedFunctional is a
TranslatedFunctional: if so the translation is combined into one translation.