WIP: fix pspace elements and shape
Power spaces now behave the same as other product
spaces. In particular, they do NOT add the shape of
their base spaces to the reported `shape`, as that
would be inconsistent with the shapes of their generated
elements.

An open question is the handling of arrays that have
a different `dtype` than `object`: regarded as sequence,
all their members are elements of the power space subspaces,
but the shape reported by the array differs from the one
of the space.