Compatibility with matplotlib-3.8.
The update methods for colorbars have been removed in favour of more global updates.

Applying the change to the `clim`s simply on the `csub` was suggested by paulhausner
in https://github.com/odlgroup/odl/pull/1635 and seems to work fine.

Actually updating the drawing of the colorbar (specifically, the ticks on it) did
not work with the fix suggested in the Matplotlib documentation, but `canvas.draw_idle()`
does seem to have the intended effect.

Fixes https://github.com/odlgroup/odl/issues/1636.