Product space unit tests compatible with modern NumPy (#1640)
* An array representation for testing product spaces with modern NumPy.

NumPy does not really do ragged arrays anymore, unless dtype=object is used - in which case however
numerical operations are largely unsupported.
This solution is a compromise. It does not work perfectly, but is usable with some special-casing
to carry out the unit tests.

* Almost-equality check that works with the new array representations for product spaces.

* Rewrite of the tests for real- and imaginary setters in product spaces.

The old ones explicitly checked on shapes (with a quite dubious special case for ℝ³).
This does not work anymore with the current reference arrays and more restrictive NumPy.
The new version uses something of a brute force attempt: we try to replicate the expected
result to match any product space structure. Not pretty, but it avoids hard-codes special
cases for specific spaces and works for the present test cases.

* Mangle code layout to satisfy pedantic linter rules.

Unfortunately reduces readability a lot, but this is how the rules are.

* Document the purpose of recursion_limit in pspace_test.