Update dependency bounds.
These correspond to the lowest version with which I was able to install through
Conda and run ODL examples. Probably even lower version would work, but
they are not readily available anymore and probably not very interesting
for users either.

I have not exhaustively checked all newer versions. The dependencies we currently
use for ODL development are considerably newer.