Don't rely on obsolete NumPy inhomogenous arrays for boundary specification.
The partition builders are quite liberal (arguably, unnecessarily) in the format of the
boolean lists specifying along which dimensions the boundary should be included in
the grid and along which not. Previously, these free-form lists were passed through
NumPy as a first step of analysing their form, but NumPy itself is now more strict
in what can be in an array (unless explicitly asked with `dtype=object`, in which
case however a plain list could be used just as well).
These checks can also be done in a more direct fashion, presuming that the caller
actually follows the specification.

Amended due to feedback by Jevgenija in https://github.com/odlgroup/odl/pull/1649#pullrequestreview-2268796987.