Manual broadcasting for the general tensor-with-inconsistent-dimensions case.
Previously, this was more or less automatically done in NumPy, but not anymore.
Normally, this is more likely to be user mistake. However, ODL actually
relies somewhat on such broadcasts when defining functions on "sparse"
mesh grids, so I added this functionality back be recursively
transversing lists of different-shape arrays (like in the old version
NumPy did, manually generating ragged dtype=object arrays).