Remove some type-inconsistent tests. (#1653)
* It is impossible to check the Astra version if none is available.

This would cause errors when running ScikitImage tests when that is the only backend available.

* Remove some type-inconsistent tests.

Reducing (aka folding) an operation like (==) over an array of e.g. floating-point
numbers does not really make any sense, because after the first application one
has a boolean which to compare to a number is ill-typed.

This only ever worked because old NumPy would apply implicit type widening, but
newer versions rightfully do not support this anymore. Since the test only
checked whether ODL gave the same results as NumPy, but NumPy does _not_ give
results for these cases anymore, it is only reasonable to remove the tests.

* Newer NumPy does not support `floor_divide` on complex arrays.

Which is reasonable; the notion of floor is dubious at best on complex numbers.