dtype=object array problems in function-on-grid evaluation (#1656)
* NumPy changed what exception is raised for ufunc-call with wrong number of arguments.

The ODL tests checked against the previously raised `ValueError`, but in newer
versions `TypeError` is raised instead, which these tests could not handle.

* Don't rely on obsolete NumPy inhomogenous arrays for boundary specification.

The partition builders are quite liberal (arguably, unnecessarily) in the format of the
boolean lists specifying along which dimensions the boundary should be included in
the grid and along which not. Previously, these free-form lists were passed through
NumPy as a first step of analysing their form, but NumPy itself is now more strict
in what can be in an array (unless explicitly asked with `dtype=object`, in which
case however a plain list could be used just as well).
These checks can also be done in a more direct fashion, presuming that the caller
actually follows the specification.

Amended due to feedback by Jevgenija in https://github.com/odlgroup/odl/pull/1649#pullrequestreview-2268796987.

* Manual broadcasting for the general tensor-with-inconsistent-dimensions case.

Previously, this was more or less automatically done in NumPy, but not anymore.
Normally, this is more likely to be user mistake. However, ODL actually
relies somewhat on such broadcasts when defining functions on "sparse"
mesh grids, so I added this functionality back be recursively
transversing lists of different-shape arrays (like in the old version
NumPy did, manually generating ragged dtype=object arrays).

* Change doc test to have floating zero.

There is a check that the result must have the correct dtype, even
if it does not need to have the correct shape.