Propose a backend-agnostic ways of checking dtype compatibility.
The implementation with a random check is an ugly hack, barely acceptable for
this purpose. Arguably, hardcoding a full matrix of what is convertible would be
a cleaner solution, but it would actuall be more problematic from a maintenance
perspective because Torch might change what conversions are supported.

The only proper solution would be to use a Torch function corresponding to
np.can_cast, but torch.can_cast does not do that, it is more restrictive.