Correct bug in Fourier post-processing.
It makes no sense to pass the `out` argument as the input parameter.
The `x` argument of the `_postprocess` method was completely unused.
Upon investigating why this never caused any problems, I found that in
all the unit tests `x is out` held true. In that case both are interchangeable,
but this cannot in general be assumed.
I can imagine no setting where it would actually be necessary to double-pass
`out` this way.

As for why the old version used `out` as the input argument: this originated in
4e4e928, where the call to `dft_postprocess_data` was refactored:
it had previously been in the `_call_pyfftw` method, where indeed the data was
stored in `out` (having come out of `pyfftw_call`). It appears that this call
was copy&pasted into the then-new `_postprocess` method, but forgotten to change
the argument to `x`.