Add an out-of-place version for NumPy "low-level" multiplication and division.
Until now, ODL always switches to in-place style in the general wrappers, but this will
change in the future because it does not play well with PyTorch.
The new SupportedNumOperationParadigms feature would in principle allow declaring that
the NumPy version supports _only_ in-place operations. That would probably work ok in
practice, with much the same behaviour as the old version, but it could lead to strange
problems when dealing with product spaces where some subspaces support only out-of-place.
Since NumPy can do out-of-place just fine, it would be pointlessly dogmatic to prevent
that.