It turns out `matplotlib-base` is not a package pip can track.
It is only a Conda that splits up matplotlib into a base and a full part,
which differ in the latter pulling in the Qt widget TK.

For pip this does not apply since is only installs Python modules anyway,
and merely uses Qt if already present on the system.

Partially reverts 7443b3c036875e23cecd4a9fd3c0015441706114.