Workaround for changed normalisation convention in PyFFT>=0.13.
In older versions of PyFFT, the `normalise_idft` argument was only
relevant for the inverse direction, else ignored. ODLs `pyfftw_call` wrapper
was modelled around this behaviour.

Starting with PyFFT-0.13, the new NumPy normalisation arguments `norm='forward'`
and `norm='backward'` (https://github.com/numpy/numpy/pull/16476) were adopted
for the high-level binders (https://github.com/pyFFTW/pyFFTW/pull/308, merged in
https://github.com/pyFFTW/pyFFTW/pull/330).
They implemented this by pre-matching the `norm` argument and setting
`normalise_idft` accordingly.
In case of the 'forward' normalisation this involves using `normalise_idft=False`
to signal that the normalisation should instead happen in the forward direction,
which was previously not even possible. In other words, in PyFFT<0.13
`normalise_idft=False` meant no normalisation at all was used (neither for fft
nor ifft, with the consequence that they weren't inverses in this case), and that
is the behaviour ODL expected.

Explicitly setting `normalise_idft=True` in this specific case restores the
desired behaviour, and is backwards-compatible because PyFFT<0.13 ignores
the `normalise_idft` argument in the forward direction.