Workaround for the normalization inconsistency with 2D parallel beam.
Astra carries out the normalization in 3D, which only matches what ODL expects in 2D
if the detector thickness is the same as one voxel in the volume. One (rather hackish)
way of ensuring this is to calculate the voxel size separately and pass it to the detector
geometry setup.