Workaround way to confirm a padding constant can be safely converted to the NumPy array type.
`can_cast` used to do a value based check, involving that e.g. `54` can be cast to `uint8` but `540` cannot.
This level of checking has been removed in NumPy 2 (https://numpy.org/neps/nep-0050-scalar-promotion.html).

It is debateable whether ODL should rely on such checks at all (after all, they do not protect from e.g. overflows
in additions, which are rather harder to debug than constant padded values), but as it stands the test suite
does require them. Attempting to wrap the value in a singleton array and checking for equality with a version
of unconstrained dtype (which NumPy automatically widens as required) is at any rate a reliable way to assess
whether the conversion has been successful in the common sense.