Generalize Fourier-sign tests over different dtypes.
The old version was fixed to single precision, but in NumPy-1 it actually always
used double under the hood.
The `all_close` condition could only be fulfilled because the double FFT had a
pointwise accuracy with relative deviations that could not be supported by
an FFT on the type that does the actual tolerance checks.
Now we explicitly select appropriate tolerances for each of the float sizes,
and we do it with `np.norm` instead of the pointwise checks (which fail for
result entries very close to zero, because a Fourier transform uses differences
between multiple different input values under the hood).