Directly use the caching decorator from functools.
This was not available in Python 2 which is why ODL provided its own decorator
which simply wrapped the functools one if-available.
As we do not support Python 2 anymore, this is now unnecessary, and the semantics
of memoization sometimes happening but not reliably is problematic.
If `cache_functions` was in some ways particularly taylored for ODL it might
make sense to keep it, but as it really just wraps the existing `lru_cache`
it is the cleanest option to use that directly.