Reliable tolerances on all tests in the tensor_test suite.
These largely weren't necessary in NumPy-1, which used double precision
for many operations even with low-precision inputs. This is not the case
anymore, and so most of these tests need to allow for a larger tolerance
when using float32 or float16. Partly this was already fixed in 93e6d6 and
249a573, but as these tests use random noise data there could still be
rare cases where the tolerance was exeeded.