Relax the power_method_opnorm test, allowing for nondeterminism.
A power method with random starting point has a small but non-vanishing chance
of giving a wrong result. This was revealed when running the tests over and
over again.
To some extend this phenomenon is inevitable (as for a general operator there
is no way to know one has found a non-maximal eigenvalue), so for now I
patched it by allowing one out of three runs of the method to give a wrong
result. It is worth discussing whether something similar should be built
into `power_method_opnorm` itself.