Resize-operator-adjoint test tolerance relative on the input magnitude.
Before, this test was relative in the results of the inner products. This usually
works, but it can fail if the inputs are almost orthogonal: in that case, both
versions of the inner product come out close to zero. `pytest.approx` requires
their discrepance to be extremely small then, which cannot be fulfilled because
the inputs (which are _not_ close to zero) incur still a fairly large floating-point
error. This lead to rare test failures (about 1:2000) in the old version.

Making the tolerance independent of the magnitude of the inner product, but instead
on the norms of the inputs, solves the problem.