Policy on what kind of weighting-arrays are allowed in the `ArrayWeighting` class.
Up until now, the NumPy-specific weighting classes allowed passing an element of
ODL spaces directly as the weighting.
This feature was not really documented, but checked in the tests.
Since the weighting classes are more low-level than the classes of spaces, they
should not be used like this. Allowing it (which the NumPy version does awkwardly
with an instance check on the specific class) is rather a matter of protecting the
user from shooting themselves in the foot than a legitimate use case. A better
safeguard is to ensure the weightings are an array that is compatible with the
specified `impl`.