Propose a new class as a single access point for the different array backends.
Most of the functionality is handled by the Python Array API, which this class defers to.
Additionally there is some meta-information, and the backends are registered so they
can always be looked up by a simple `impl` string as customary.

I found it easiest to formulate this as a dataclass, which is concise and clear yet allows
for the central registry. Concrete backends are simply instances of the `ArrayBackend` class.
An alternative would be to let them be subclasses. This is not really necessary, but would
arguably make the syntax less crammed and allows for better documentation.
It would be less appropriate in the sense that those subclasses would be _singleton classes_,
and would need to be registered in a different way. Python does allow this (via decorators
or metaclasses), but that would be much more arcane.