Move some functionality from `TensorSpace` into `ArrayBackend`.
These are properties that do not pertain to a specific space with its
mathematical meaning, but only to the computational backend that it
happens to use for storing its internal data.

This makes some use cases slightly more verbose, but I would say it is
justified by the simplification of the interface to the anyways rather
bloated `TensorSpace` class.

In principle, the `array_namespace` method could also be removed from
`TensorSpace`, but this is invoked so often that the cost/benefit ratio
is less favourable.