Turn `default_dtype` into a standalone function.
This function does not do anything specific to a particular space, but is
rather a property of the array backend. It is also useful without the
context of an already defined space, indeed it can be particularly useful
to find out how to initialize such a space in the first place.