Change the default dtype back to double-precision.
This is how it was in old ODL. There is a discussion to be had about this,
since single precision does give much better performance on GPUs. The lower
precision is unlikely to matter in the typical applications, though it could
still trip up users particularly in purer maths contexts.