Make division-by-zero-behaviour consistent with ODL-0.8.
Old ODL implemented, somewhat accidentally, a signalling-infinities behaviour.
Dividing an ODL object by a scalar zero would raise an exception, not because
it was explicitly checked but because it was under the hood implemented by
a `_lincomb` call multiplying by `1.0 / scalar`. That simple Python division
was what raised the exception.
Note that no exception is raised when dividing by an ODL object that contains
zero entries. Generally, the array backends do not error when a zero division
happens (though they may warn). Instead the default is to rely on the IEEE-754
infinity values for floating-point zero division. Arguably ODL should just
follow this convention.