Add an explicit-disabled implementation of `__array__`.
This will be the first thing NumPy tries to call when implicitly converting to
a raw array (which we do not want to happen, as it would lead to inconsistent
behaviour). Without this in place, NumPy would hack its way around by using
`getitem`, which is _both_ inconsistent and also very slow.