Implement the elementwise inequality operators on `LinearSpace`.
The semantics of these are somewhat confusing, becase the `==` operator gives
as single bool whereas the array API specifies that all the [in]equality ops
give an array.
For now, we yield an ODL boolean element from the _in_equality operators, but
a single boolean from the `==` and `!=`. This may need to be reconsidered in
the future.