Implement `writable_array` in a disciplined, not NumPy-specific manner.
The use case of this context manager is to modify the data underlying general ODL
objects. It is further used to ensure the modifying routines get to see the array
in contiguous form. Previously, this relied on passing keyword-arguments on to
a `np.asarray` call, which is obviously not valid for arbitrary API-backends, particularly
not the `order=c` combination used for achieving contiguity.

Now, there is an explicitly named parameter specificly for ensuring contiguity.
The old version also allowed e.g. specifying a different `dtype`. I do not think
this is a good idea, as changing the dtype here would _always_ cause loss of
precision (and/or imaginary parts), either during the initial conversion or when
loading the data back into the original object.