Update Astra tomo backend to changes made to make ODL ArrayAPI compatible.
I changed the signature of `astra_data`. Previously it accepted either raw arrays
or ODL space-elements as the data. This complicates the logic, and I feel it
is unnecessary since `astra_data` is a fairly low-level function that anyways
just used the ODL object as an array. So now it should always be an array (in
the future it should however be possible to pass non-NumPy arrays).