Make Astra-Cuda independent of PyTorch dependency.
Although it is sensible to use Torch in the Cuda case to avoid copies, this is not a necessity.
The module currently used the `torch` import only for a type signature, and as type signatures
anyway don't really do anything in Python this can easily dispensed with to make ODL compile
also without PyTorch installed.