Make `_convert_dtype` more flexible and efficient.
The previous version awkwardly looped through `TENSOR_SPACE_IMPLS` (though this
function does not really have anything to do with spaces, only low-level arrays),
and then needed to look up an array backend in a separate step. It makes more
sense to loop through the registered array backends right away. This makes it
also possible to provide a custom selection of backends, which is useful when
looking up dtypes for only a specific backend.