Omit dtype/impl/device from space's __repr__ if they are the default values.
This restores the old ODL behaviour, corresponds to how NumPy shows array, and makes simple experiments in the REPL
more beginner-friendly.
N.B.: if the device or impl is non-default, the dtype is also shown even if that is default.
This is necessary to retain order of positional arguments. An alternative would be to
have impl as a keyword argument in that case, but that would make for almost the same
verbosity as an unnecessary dtype shown.