Use the ODL version of `abs` iff dealing with ODL objects.
Calling `odl.abs` on plain Python scalars is currently not supported, and likely
will not be (e.g. Torch does not allow it either).

It used to be possible to directly call `np.abs` on ODL space-elements.
This was nice in the sense that it worked on anything from Python
scalars over NumPy arrays to ODL, but only thanks to the various
problematic case distinction and ufuncs tricks. These are not supported
anymore because they were in the way of backend-agnosticism.