Correct the logic of when the weighting of a `DiscretizedSpace` needs to be printed.
The purpose of this check is to ensure an explicit weighting is shown when the
default volume calculation does not make sense. This is the case for integral types,
thus the check on `is_floating_dtype`. In d48372a, this was accidentally changed
to `is_real_floating_dtype`, with the result that complex `DiscretizedSpace`s then
ended up printing the weighting even when it equalled the cell volume.

How this happened: d48372a was only meant to roll back changes in 42fc5ca, which
had replaced `is_real_floating_dtype` with `is_floating_dtype` in several places.
In `discr_space.py` however, `is_real_floating_dtype` had never been used anyway,
so the "rollback" to it in d48372a was incorrect.