Use the zero-safe xlogy operation (from SciPy) for KL-divergence.
This is what ODL originally did too (albeit NumPy-only). It was changed to the
manual `x * log(y)`, which however `NaN`s if there are any zeroes in `y` (in
this case, in the prior).

Emilien already came up with a way of using special SciPy functions on ODL objects,
but so far it was not used in `KullbackLeibler`.