Make the `_call` of `MatrixOperator` exclusively out-of-place.
The Array API does not offer a matmul or dot operation with an out argument,
so an in-place call can only be achieved with the NumPy-specific `dot` (short
of using low-level BLAS libraries).
The disadvantage of memory allocation for out-of-place is arguably a minor
concern for `MatrixOperator`, since the matrix occupies 𝑂(𝑁²) space but the
output only 𝑂(𝑁). This argument would not apply for sparse matrices, but those
used an out-of-place call even before, as SciPy does not offer an in-place one.