Generalize `check_device` to accept not only plain string identifiers but also e.g. torch.device objects.
Although we prefer to identify devices by strings (which is backend-agnostic),
the backends themselves usually do not do it this way - e.g. Torch has a
specific type for identifying devices. It it fairly natural to use such an
identifier when initializing an ODL tensor space (for example, to match the
device of some give Torch tensor input).