Consistent version of the current semantics for `__eq__`.
The `shape` etc. checks were redundant as that is already covered by membership
in the space.
The type of `array_namespace.all` would depend on the backend, which is
undesirable here.

N.B. all of this may be moot as the semantics of checking whether the two
arguments are equal is _not what the array API expects_ for the `==` operator.
Instead it prescribes the Matlab-like behaviour of giving a new array back,
with pointwise comparisons.
Discussion in https://github.com/odlgroup/odl/issues/1695.