Ensure `set_zero` leaves only zeroes, even if NaNs were previously present.
This was the cause of a problem @Emvlt and me spent considerable time hunting down: some
proximal calls would rarely indeterministically give NaN results.
The problem turned out to be that `set_zero` does not tolerate if the input contains
`NaN` values - a consequence of the (anyways somewhat whimsical) implementation in terms
of `lincomb`. N.B. is _not_ because `lincomb` works incorrectly: it only follows the
IEEE-754 convention that NaNs be poisonous.
Rather, the problem is that `set_zero` applied it to memory that was in some use cases
completely uninitialized, and could thus contain NaN in some entries.

For some reason, this never occured in NumPy, but it does in PyTorch.

The fix is to implement `set_zero` in a way that directly writes to the array, completely
ignoring any previous content.