Revert changes in 8509070, superseded by de365f2a.
Emilien was on the right track that the problem was related to `lincomb`, but
it was its inappropriate use in `set_zero`. The use in the proximal's `call`
method is fine, because `self.lincomb` (`self: LinearSpaceElemen`) does not
use `self` as part of the input, only as the writing destination.
In this role, it is safe and has the advantage of memory reuse over the
`out[:] = ...` alternative.