Rename the device-changing method to be in line with the array API.
This means it is not in line with `astype` anymore, which has however anyway a
rather different purpose (it changes the actual numerics), so this is perhaps
not a bad thing.

`to_device` also has the advantage that there can be a corresponding `to_impl`,
which is obvious to read whereas `asimpl` would be easily misparsed "a-simpl".