Attempt at making the sampling functions in discr_utils PyTorch-compatible.
It seems to work for simple sampling of a scalar valued function on 1D domain, not
sure about general cases yet.

I removed the ability to have an out argument, as that involved a lot of complexity
for very little gain (performance should not matter at this point, because sampling
usually happens only at the start of an iterative process, and it is slow anway
because data is copied to the backend).