Properly support non-NumPy backends as the displacement of interpolation methods.
This was previously attempted in 43e56ec, but with some problems due to which I rolled
it back in 1a7ce80. It was however wrong to think that other backends would not be useful,
because the displacement field of a deformation will in fact typically be a full 3D array.
The interesting parts of the interpolation did also already work on PyTorch in @Emvlt's
version, it was only a matter of initializing them suitably.

This version now works in most situations; the tests that still fail are only
1. interpolation on an array of characters. IMO this is not something ODL should support at all.
2. nearest-neighbour interpolation at the midpoint between two nodes. Here there is a discrepancy
between NumPy and PyTorch, apparently in `searchsorted`. Not sure about that one, but arguably
the nearest neighbour is undefined in that situation and so the test should not make an assumption
either way.

Far more important is that _linear_ interpolation works, which it seems to do.