Do not attempt to change the array type in the `__init__` method of the backend-specific tensor classes.
This is already taken care for by the base tensor class' `element` method.
Doing it again in the `__init__` method is just an opportunity for unnecessary copies
to happen, which can have a serious performance impact since `element` is called so often.
Also, calling the `array_constructor` turns out to disrupt PyTorch autograd.

The only downside of not doing the conversion is that the classes by themselves
cannot be instantiated with data from other backends. This is mostly irrelevant though,
because you would only use e.g. `NumpyTensor` directly if you are sure you are dealing
with NumPy data. In fact, users should not do that at all but always go through the
`.element` method instead.