Workaround for multiple-initialization of `ArrayBackend`s when running PyTest on the modules.
`ArrayBackend` is designed to allow every backend to be registered only once (otherwise there would
be problems with nonunique helper functions).
In a normal import scenario, this is fine, but when running `pytest` on the entirety of ODL the
definition ends up being executed once for some import and then again when testing the module
itself. This kludge catches the exception raised in that specific situation.