Fix the PyShearlab operators (for ODL-1.0 with numpy impl).
This is still somewhat scrappy. For one thing, PyShearlab has not been
maintained for 8 years and can only be used in a patched version with
modern NumPy (https://github.com/JanVictor-Otte/pyShearLab-MIND/).
The NumPy-only restriction is justified in this case, since PyShearlab
is a CPU-only library.

It would be good to replace it with something more up to date.