Add (empty) `__init__` files to all directories part of the ODL hierarchy.
These directories were added in https://github.com/odlgroup/odl/pull/1697.
When testing them locally, Python had no problems finding the modules even
without an `__init__` file (in these director, no module-remapping was needed),
but as it turns out at least an empty `__init__` is necessary for the members
of a directory to be part of a PyPI release.

I consider to explicitly list the directory contents as imports in those `__init__`
files and adding them to `__all__`, but that seems to correspond to the default
behaviour with an empty file. The latter is preferable since it avoids duplication
of information (which might go out of sync later on).