Removal of an __array_priority__ attribute from the LinearSpaceElement class.
This was a leftover from the moment where the order of operation was delegated to NumPy.
I believe we forgot to remove it and in some corner cases ODL still relied on this to resolve the order in which applying the operation.
When i went through the code, i also found that the way we extract the precision type of a np dtype (lets say the float32 part of np.float32) with str(dtype) was not strong enough. Indeed, when doing str(np.float32), we got a <class np.float32> string which is definitely not a good thing. To do so, I replaced this function in the NumpyArrayBackend. All the np tests pass but we're never too sure :,)